+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log
 
+## [Version 7.0.0](https://github.com/yannickl/QRCodeReader.swift/releases/tag/7.0.0)
+Released on 2016-06-15.
+
+- Swift 3 supports
+
+`QRCodeReader`:
+- Use `didFindCode` instead of `didFindCodeBlock`
+- Use `isRunning` instead of `running`
+- Use `videoOrientation` instead of `videoOrientationFromDeviceOrientation`
+- `hasFrontDevice` is a property
+- `isTorchAvailable` is a property
+
 ## [Version 6.0.0](https://github.com/yannickl/QRCodeReader.swift/releases/tag/6.0.0)
 Released on 2016-03-22.